data/requirements.txt
boto3>=1.26.0
pandas>=1.5.0
pyarrow>=10.0.0
pyiceberg>=0.4.0
google-cloud-bigquery>=3.15.0
